requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
line-bot-sdk==3.7.0
google-generativeai==0.3.2
supabase==2.3.7
python-dotenv==1.0.1
httpx==0.26.0
python-multipart==0.0.9
jinja2==3.1.3
